fixed the names
