+++ backend/src/server.ts
@@ -1,39 +1,17 @@
-import app from './index'
-import { Server, Socket } from 'socket.io'
-import socketHandlers from './sockets/sockets'
-import { User } from './interfaces'
-import memoryStorage from './sockets/sessionStorage'
-import clustering from './clusters/clusters'
+import express, { Request, Response } from 'express'
+import http from 'http'
 
-const port = process.env.PORT || 5000
-const io = new Server(app, {
-    cors: {
-        origin: "http://localhost:8080",
-    }
+const app = express()
+
+// middleware
+app.use(express.json())
+app.use(express.urlencoded({ extended: false }))
+
+// routes
+app.get('/ping', (_: Request, res: Response) => {
+  res.status(200).json({ message: 'pong' })
 })
 
-const DB : Array<User> = [] 
-const sessionDB = new memoryStorage()
+const server = http.createServer(app)
 
-io.use((socket: Socket, next: any) => {
-  socketHandlers.middleware(socket, next, sessionDB)
-})
-
-io.on('connection', (socket: Socket) => {
-	socketHandlers.main(socket, io, DB, sessionDB)
-})
-
-if (process.env.NODE_ENV === 'production') {
-  console.log('Production Server Online')
-  clustering(() => {
-    app.listen(port, () => {
-      console.log('🚀 Server Starting At http://localhost:' + port)
-    })
-  })
-} else {
-  console.log('Development Server Online')
-  app.listen(port, () => {
-    console.log('🚀 Server Starting At http://localhost:' + port)
-  })
-}
-
+export default server
