+++ backend/tests/server.test.ts
@@ -1,5 +1,5 @@
 import request from 'supertest'
-import app from '../src/index'
+import app from '../src/server'
 
 // testing if express works
 describe('Ping Server', () => {
